spenny_trail5s_fully_updated/spenny_trail5s/requirements.txt
cat > requirements.txt <<EOF
pandas
yfinance
python-dotenv
dropbox
openai
pdfplumber
python-docx
pytesseract
Pillow
requests
EOF
